requirements.txt
absl-py==1.4.0
accelerate==1.10.0
aiofiles==23.1.0
altair==4.2.2
annotated-types==0.7.0
anyio==3.6.2
anywidget==0.9.18
astor==0.8.1
asttokens==3.0.0
attrdict==2.0.1
Babel==2.12.1
bbox_visualizer==0.2.2
bce-python-sdk==0.8.83
blinker==1.6.2
cachetools==5.3.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==3.4.3
click==8.1.3
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cssselect==1.2.0
cssutils==2.6.0
cycler==0.10.0
Cython==3.1.3
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.13
filelock==3.19.1
fire==0.5.0
Flask==2.3.1
flask-babel==3.1.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2025.7.0
ftfy==6.3.1
future==0.18.3
gevent==22.10.2
googleapis-common-protos==1.59.0
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
huggingface-hub==0.34.4
idna==2.10
imgaug==0.4.0
importlib_resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.15
kiwisolver==1.3.1
linkify-it-py==2.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
oauthlib==3.2.2
open_clip_torch==3.1.0
opencv-python==4.12.0.88
openpyxl==3.1.2
packaging==25.0
paddleocr==2.6.1.3
paddlepaddle==2.2.2
pandas==2.3.1
parso==0.8.4
pdf2docx==0.5.6
peft==0.17.0
piexif==1.1.3
pillow==11.3.0
polygraphy==0.49.26
premailer==3.10.0
prompt_toolkit==3.0.51
psutil==7.0.0
psygnal==0.14.1
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylabel==0.1.55
pyparsing==2.4.7
pytesseract==0.3.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rf100vl==1.0.0
rfdetr==1.2.1
roboflow==0.2.4
rsa==4.7.2
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.12.2
sentry-sdk==1.15.0
six==1.17.0
stack-data==0.6.3
supervision==0.26.1
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
torch==2.7.0
torchvision==0.22.0
tqdm==4.64.0
traitlets==5.14.3
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
urllib3==1.26.6
uvicorn==0.21.1
visualdl==2.5.1
wcwidth==0.2.13
wget==3.2
widgetsnbextension==4.0.14
x2paddle==1.4.1
zipp==3.23.0
